Works with 3.8 also, still need to fix muzzle again
