+++ dd-java-agent/instrumentation/netty-3.8/src/main/java/datadog/trace/instrumentation/netty38/client/NettyHttpClientDecorator.java
@@ -1,21 +1,17 @@
-package datadog.trace.instrumentation.netty39.server;
+package datadog.trace.instrumentation.netty38.client;
 
 import static org.jboss.netty.handler.codec.http.HttpHeaders.Names.HOST;
 
-import datadog.trace.bootstrap.instrumentation.decorator.HttpServerDecorator;
-import java.net.InetSocketAddress;
-import java.net.SocketAddress;
+import datadog.trace.bootstrap.instrumentation.decorator.HttpClientDecorator;
 import java.net.URI;
 import java.net.URISyntaxException;
 import lombok.extern.slf4j.Slf4j;
-import org.jboss.netty.channel.Channel;
 import org.jboss.netty.handler.codec.http.HttpRequest;
 import org.jboss.netty.handler.codec.http.HttpResponse;
 
 @Slf4j
-public class NettyHttpServerDecorator
-    extends HttpServerDecorator<HttpRequest, Channel, HttpResponse> {
-  public static final NettyHttpServerDecorator DECORATE = new NettyHttpServerDecorator();
+public class NettyHttpClientDecorator extends HttpClientDecorator<HttpRequest, HttpResponse> {
+  public static final NettyHttpClientDecorator DECORATE = new NettyHttpClientDecorator();
 
   @Override
   protected String[] instrumentationNames() {
@@ -24,7 +20,7 @@ public class NettyHttpServerDecorator
 
   @Override
   protected String component() {
-    return "netty";
+    return "netty-client";
   }
 
   @Override
@@ -40,24 +36,6 @@ public class NettyHttpServerDecorator
     } else {
       return uri;
     }
-  }
-
-  @Override
-  protected String peerHostIP(final Channel channel) {
-    final SocketAddress socketAddress = channel.getRemoteAddress();
-    if (socketAddress instanceof InetSocketAddress) {
-      return ((InetSocketAddress) socketAddress).getAddress().getHostAddress();
-    }
-    return null;
-  }
-
-  @Override
-  protected Integer peerPort(final Channel channel) {
-    final SocketAddress socketAddress = channel.getRemoteAddress();
-    if (socketAddress instanceof InetSocketAddress) {
-      return ((InetSocketAddress) socketAddress).getPort();
-    }
-    return null;
   }
 
   @Override
