+++ dd-java-agent/instrumentation/netty-3.8/src/main/java/datadog/trace/instrumentation/netty38/server/HttpServerTracingHandler.java
@@ -1,8 +1,8 @@
-package datadog.trace.instrumentation.netty39.server;
+package datadog.trace.instrumentation.netty38.server;
 
 import datadog.trace.bootstrap.ContextStore;
-import datadog.trace.instrumentation.netty39.ChannelTraceContext;
-import datadog.trace.instrumentation.netty39.util.CombinedSimpleChannelHandler;
+import datadog.trace.instrumentation.netty38.ChannelTraceContext;
+import datadog.trace.instrumentation.netty38.util.CombinedSimpleChannelHandler;
 import org.jboss.netty.channel.Channel;
 
 public class HttpServerTracingHandler
