+++ dd-java-agent/instrumentation/netty-3.8/src/main/java/datadog/trace/instrumentation/netty38/server/NettyHttpServerDecorator.java
@@ -1,17 +1,21 @@
-package datadog.trace.instrumentation.netty39.client;
+package datadog.trace.instrumentation.netty38.server;
 
 import static org.jboss.netty.handler.codec.http.HttpHeaders.Names.HOST;
 
-import datadog.trace.bootstrap.instrumentation.decorator.HttpClientDecorator;
+import datadog.trace.bootstrap.instrumentation.decorator.HttpServerDecorator;
+import java.net.InetSocketAddress;
+import java.net.SocketAddress;
 import java.net.URI;
 import java.net.URISyntaxException;
 import lombok.extern.slf4j.Slf4j;
+import org.jboss.netty.channel.Channel;
 import org.jboss.netty.handler.codec.http.HttpRequest;
 import org.jboss.netty.handler.codec.http.HttpResponse;
 
 @Slf4j
-public class NettyHttpClientDecorator extends HttpClientDecorator<HttpRequest, HttpResponse> {
-  public static final NettyHttpClientDecorator DECORATE = new NettyHttpClientDecorator();
+public class NettyHttpServerDecorator
+    extends HttpServerDecorator<HttpRequest, Channel, HttpResponse> {
+  public static final NettyHttpServerDecorator DECORATE = new NettyHttpServerDecorator();
 
   @Override
   protected String[] instrumentationNames() {
@@ -20,7 +24,7 @@ public class NettyHttpClientDecorator extends HttpClientDecorator<HttpRequest, H
 
   @Override
   protected String component() {
-    return "netty-client";
+    return "netty";
   }
 
   @Override
@@ -36,6 +40,24 @@ public class NettyHttpClientDecorator extends HttpClientDecorator<HttpRequest, H
     } else {
       return uri;
     }
+  }
+
+  @Override
+  protected String peerHostIP(final Channel channel) {
+    final SocketAddress socketAddress = channel.getRemoteAddress();
+    if (socketAddress instanceof InetSocketAddress) {
+      return ((InetSocketAddress) socketAddress).getAddress().getHostAddress();
+    }
+    return null;
+  }
+
+  @Override
+  protected Integer peerPort(final Channel channel) {
+    final SocketAddress socketAddress = channel.getRemoteAddress();
+    if (socketAddress instanceof InetSocketAddress) {
+      return ((InetSocketAddress) socketAddress).getPort();
+    }
+    return null;
   }
 
   @Override
