+++ dd-java-agent/instrumentation/netty-3.8/src/main/java/datadog/trace/instrumentation/netty38/util/CombinedSimpleChannelHandler.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.netty39.util;
+package datadog.trace.instrumentation.netty38.util;
 
 import org.jboss.netty.channel.ChannelEvent;
 import org.jboss.netty.channel.ChannelHandlerContext;
